+++ src/components/base/Popover/Button.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export default styled.button`
+    background: transparent;
+    border: none;
+    color: inherit;
+    cursor: inherit;
+    outline: 0;
+    cursor: pointer;
+    &:hover {
+        color: #ff8c31;
+    }
+`;
