+++ src/components/base/Popover/List/Item.tsx
@@ -1,0 +1,6 @@
+
+import styled from 'styled-components';
+export default styled.li`
+    list-style-type: none;
+    white-space: nowrap;
+`;
