+++ src/components/base/Popover/List/index.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export default styled.ul`
+    margin: 0;
+    padding: 0;
+
+    height: 100px;
+    width: 100px;
+`;
